Add next tick delay in testPortAsync
For some reason, on linux, the port is not released in the same tick, thus if you immediately try to create a server on the returned port from this module, the port is not allocated successfully and Node doesn't throw an error. By adding a single tick delay, we avoid this problem.